requirements.txt
a2wsgi==1.8.0
anyio==4.0.0
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.1
exceptiongroup==1.1.3
flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
python-multipart==0.0.6
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
sniffio==1.3.0
starlette==0.32.0.post1
typing-extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
werkzeug==3.0.1
zipp==3.17.0
